Configure Dockerfiles

+++ WebsiteResume.Domain/Entities/Skill.cs
@@ -3,6 +3,6 @@
     public class Skill
     {
         public string Id { get; set; }
-        public short Order { get; set; }
+        public byte Order { get; set; }
     }
 }
